Create an svg Flag Icon Tag

Create an html <img> tag for a country flag svg file that is shipped with the package. The flag icons are by Lipis and were downloaded from https://github.com/lipis/flag-icons.

Usage

flag_icon(code)

Arguments

code

character with the flag code, e.g., "ch" or "de". To find available flag codes, use get_flag_info().